Branching for 1.7 release.

+++ core/10.5/fusefs/fuse_knote.c
@@ -118,7 +118,7 @@ filt_fusevnode(struct knote *kn, long hint)
     }
 
     if (kn->kn_sfflags & hint) {
-        kn->kn_fflags |= (int)hint;
+        kn->kn_fflags |= hint;
     }
 
     if ((hint == NOTE_REVOKE)) {
